requirements.txt
# AgentEconomist Core Dependencies
# 
# Python Environment Requirement: Python 3.11+ (required by LangGraph CLI)
# 
# Installation Steps (Conda recommended):
# 
# === Method 1: Conda (Recommended, Default) ===
# 1. Create conda environment:
#    conda create -n economist python=3.11
#    conda activate economist
#
# 2. Install PyTorch (CPU version):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#    Or use Aliyun mirror: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu -i https://mirrors.aliyun.com/pypi/simple/
#
# 3. Install other dependencies:
#    pip install -r requirements.txt
#
# === Method 2: venv (Alternative) ===
# 1. Create virtual environment:
#    python3.11 -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    # or venv\Scripts\activate  # Windows
#
# 2. Install PyTorch (CPU version):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# 3. Install other dependencies:
#    pip install -r requirements.txt

# Basic utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Distributed computing
ray>=2.0.0

# Vector database
qdrant-client>=1.16.0  

# Machine Learning & Natural Language Processing
transformers>=4.35.0
sentence-transformers>=2.2.0
adapters>=0.2.0  # For SPECTER2 adapters (Database module)
# torch>=2.0.0  # Install CPU version separately in install.sh

# LangChain & LangGraph (AgentEconomist core modules)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-sdk>=0.1.30
langgraph-cli[inmem]>=0.4.0  # LangGraph CLI with in-memory server (requires Python 3.11+)

# Web server (for start_server.py)
fastapi>=0.100.0
uvicorn>=0.20.0

# HTTP client
httpx>=0.27.0
